requirements.txt
absl-py==1.4.0
addict==2.4.0
blessed==1.20.0
bytecode==0.15.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
easydict==1.10
einops==0.6.1
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
future==0.18.3
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gpustat==1.1
grpcio==1.56.0
h5py==3.10.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
json-tricks==3.17.1
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
mmcv-full==1.6.0
model-index==0.1.11
motmetrics==1.4.0
munkres==1.1.4
nvidia-ml-py==11.525.112
oauthlib==3.2.2
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
Pillow==9.4.0
platformdirs==3.8.0
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
shapely==2.0.1
sheen==0.1.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
thop==0.1.1.post2209072238
timm==0.9.2
tomli==2.0.1
torch==1.10.2
torch-dct==0.1.6
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.11.3
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
visualizer==0.0.1
wcwidth==0.2.6
Werkzeug==2.3.6
xmltodict==0.13.0
xtcocotools==1.13
yapf==0.40.1
zipp==3.15.0
